+++ CHANGELOG.md
@@ -3,3 +3,5 @@
 
 - [Feature: Support custom rate limit reject response info](https://github.com/Tencent/spring-cloud-tencent/pull/128)
 - [Feature: Optimize config server address](https://github.com/Tencent/spring-cloud-tencent/pull/130)
+- [Feature: Remove spring-javaformat-maven-plugin](https://github.com/Tencent/spring-cloud-tencent/pull/131)
+- [feat:refactor loadbalancer module as a basic module for router and circuit breaker.](https://github.com/Tencent/spring-cloud-tencent/pull/136)